[CI] move tests to website_event_attendee_fields
-- it's that module that responsible for it

[CI] updates tests accordting to latest update user do can update contact, but
only hisown
