+++ website_event_attendee_fields/tests/__init__.py
@@ -1,2 +1,3 @@
 from . import test_backend
+from . import test_security
 from . import test_workflow
