+++ website_event_attendee_fields/tests/test_backend.py
@@ -66,4 +66,3 @@ class TestBackend(common.TestCase):
         with self.assertRaises(AssertionError):
             obj = WebsiteEventControllerExtended()
             obj.registration_confirm(event, **post)
-
